src/ai/chat_snowflake_test/requirements.txt
python-dotenv
cryptography
snowflake-snowpark-python
numpy
langgraph
langchain-snowflake==0.2.2